pyproject.toml
[tool.poetry]
name = "mcp-server-needle"
version = "0.1.0"
description = "MCP server for Needle RAG integration"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
mcp-core = "^0.1.0"
needle-python = "^0.4.0"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 